+++ src/components/directives/image/test/unit/main.ts
@@ -6,6 +6,8 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+import type { Page } from 'playwright';
+
 import test from 'tests/config/unit/test';
 
 import {
@@ -73,6 +75,15 @@ test.describe('components/directives/image', () => {
 	);
 
 	test(
+		"the created `<img>` element should always have the `src` attribute, even if it's not explicitly provided",
+
+		async ({page}) => {
+			const {image} = await renderDirective(page, {});
+			await test.expect(image.getAttribute('src')).toBeResolvedTo('');
+		}
+	);
+
+	test(
 		[
 			'if the `sources` parameter is specified, ' +
 			'the directive should be rendered using the `<picture>` element instead of `<img>`'
@@ -119,17 +130,23 @@ test.describe('components/directives/image', () => {
 		await test.expect(image.getAttribute('data-on-load-called')).toBeResolvedTo('1');
 	});
 
-	test('the provided `onError` handler should be called upon image loading errors', async ({page}) => {
-		const {image} = await renderDirective(page, {
-			src: BROKEN_PICTURE_SRC,
+	test.describe('the provided `onError` handler should be called upon image loading errors', () => {
+		test('if the specified `src` cannot be loaded', ({page}) => checkOnErrorHandler(page, BROKEN_PICTURE_SRC));
 
-			onError: (el: Element) => {
-				el.setAttribute('data-on-error-called', '1');
-			}
-		});
+		test('if the `src` attribute is not explicitly provided', ({page}) => checkOnErrorHandler(page));
 
-		await waitForAttribute(page, image, 'data-on-error-called');
-		await test.expect(image.getAttribute('data-on-error-called')).toBeResolvedTo('1');
+		async function checkOnErrorHandler(page: Page, src?: string): Promise<void> {
+			const {image} = await renderDirective(page, {
+				src,
+
+				onError: (el: Element) => {
+					el.setAttribute('data-on-error-called', '1');
+				}
+			});
+
+			await waitForAttribute(page, image, 'data-on-error-called');
+			await test.expect(image.getAttribute('data-on-error-called')).toBeResolvedTo('1');
+		}
 	});
 
 	test(