+++ src/components/form/b-button/b-button.ts
@@ -204,10 +204,6 @@ class bButton extends iButtonProps implements iOpenToggle, iVisible, iWidth, iSi
 	 * @emits `click(e: Event)`
 	 */
 	protected async onClick(e: Event): Promise<void> {
-		if (this.type !== 'link') {
-			e.stopPropagation();
-		}
-
 		switch (this.type) {
 			case 'link':
 				break;