+++ src/lang/index.ts
@@ -26,6 +26,7 @@ try {
 			};
 		});
 	});
+
 } catch {}
 
 export default langPacs;