+++ src/components/directives/image/CHANGELOG.md
@@ -9,9 +9,17 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## 4.0.0-beta.84 (2024-04-11)
+
 #### :bug: Bug Fix
 
-* Fixed the preview show time for cached images
+* Fixed an error with incorrect handling of the empty required `src` attribute
+
+## 4.0.0-beta.81 (2024-04-01)
+
+#### :bug: Bug Fix
+
+* Fixed an error with incorrect handling of errors caused by the Async wrapper
 
 ## v4.0.0-beta.47 (2024-01-16)