+++ src/components/friends/async-render/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## 4.0.0-beta.81 (2024-04-01)
+
+#### :bug: Bug Fix
+
+* Fixed rendering of nested asynchronous tasks
+
 ## v4.0.0-beta.67 (2024-02-26)
 
 #### :bug: Bug Fix